Moved module resolution test in mods.rs
